remove headers, moved folders for clarity
