+++ app/src/main/java/com/realityexpander/tasky/data/repository/authRepositoryImpls/AuthRepositoryFakeImpl.kt
@@ -1,11 +1,11 @@
-package com.realityexpander.tasky.data.repository.remote.authRepositoryImpls
+package com.realityexpander.tasky.data.repository.authRepositoryImpls
 
 import com.realityexpander.tasky.common.Email
 import com.realityexpander.tasky.common.Exceptions
 import com.realityexpander.tasky.common.Password
 import com.realityexpander.tasky.common.Username
-import com.realityexpander.tasky.data.convertersDTOEntityDomain.toDomain
-import com.realityexpander.tasky.data.repository.AuthInfoDTO
+import com.realityexpander.tasky.data.common.convertersDTOEntityDomain.toDomain
+import com.realityexpander.tasky.data.repository.remote.AuthInfoDTO
 import com.realityexpander.tasky.data.repository.remote.IAuthApi
 import com.realityexpander.tasky.data.repository.local.IAuthDao
 import com.realityexpander.tasky.domain.AuthInfo
