+++ app/src/main/java/com/realityexpander/tasky/data/repository/remote/authApiImpls/TaskyApi.kt
@@ -2,7 +2,7 @@ package com.realityexpander.tasky.data.repository.remote.authApiImpls
 
 import com.realityexpander.tasky.BuildConfig
 import com.realityexpander.tasky.data.repository.remote.ApiCredentialsDTO
-import com.realityexpander.tasky.data.repository.AuthInfoDTO
+import com.realityexpander.tasky.data.repository.remote.AuthInfoDTO
 import retrofit2.Response
 import retrofit2.http.*
 
@@ -11,13 +11,11 @@ interface TaskyApi {
     @POST("login")
     suspend fun login(
         @Body credentials: ApiCredentialsDTO,
-        //@Header("x-api-key") apiKey: String = API_KEY
     ): Response<AuthInfoDTO>
 
     @POST("register")
     suspend fun register(
         @Body credentials: ApiCredentialsDTO,
-        //@Header("x-api-key") apiKey: String = API_KEY
     ): Response<Void>
