+++ app/src/main/java/com/realityexpander/tasky/di/RepositoryModule.kt
@@ -1,6 +1,6 @@
 package com.realityexpander.tasky.di
 
-import com.realityexpander.tasky.data.repository.remote.authRepositoryImpls.AuthRepositoryImpl
+import com.realityexpander.tasky.data.repository.authRepositoryImpls.AuthRepositoryImpl
 import com.realityexpander.tasky.data.repository.local.AuthDaoFakeImpl
 import com.realityexpander.tasky.data.repository.local.IAuthDao
 import com.realityexpander.tasky.data.repository.remote.authApiImpls.AuthApiImpl
